Better readme and rails 3 routing
The readme has the model and user mdoel names backwards.
I've updated the readme - but probably the actual generators should be
fixed.

Also - I've added proper Rails 3 routing - because the alternative simply
wasn't working with rails 3.
We can probably add some sort of "if RAILS_VERSION == 3" check in here and
use the appropriate one depending on the current rails version in use.